# api-docs.json
{"swagger":"2.0","info":{"version":"1.0","title":"http://localhost:5000"},"basePath":"/","paths":{"":{"get":{"operationId":"","parameters":[],"responses":{"0":{},"200":{},"500":{}}}},"/accounts":{"get":{"operationId":"accounts","parameters":[{"name":"account_id","in":"query","required":false,"type":"string"},{"name":"customer_id","in":"query","required":false,"type":"string"},{"name":"alias","in":"query","required":false,"type":"string"}],"responses":{"0":{},"200":{}}}},"/credit_cards":{"get":{"operationId":"credit_cards","parameters":[{"name":"number","in":"query","required":false,"type":"string"},{"name":"alias","in":"query","required":false,"type":"string"},{"name":"customer_id","in":"query","required":false,"type":"string"},{"name":"brand","in":"query","required":false,"type":"string"}],"responses":{"0":{},"200":{}}}},"/credit_cards/statement":{"get":{"operationId":"statement","parameters":[{"name":"number","in":"query","required":false,"type":"string"},{"name":"alias","in":"query","required":false,"type":"string"},{"name":"customer_id","in":"query","required":false,"type":"string"},{"name":"brand","in":"query","required":false,"type":"string"}],"responses":{"0":{},"200":{},"500":{},"404":{}}}},"/customers":{"get":{"operationId":"customers","parameters":[],"responses":{"200":{},"500":{}}}},"/customers/register":{"post":{"operationId":"","consumes":["application/json"],"parameters":[{"in":"body","name":"body","description":"Request body","required":true}],"responses":{"0":{},"200":{},"500":{},"404":{}}}},"/fill":{"get":{"operationId":"fill","parameters":[],"responses":{"0":{},"200":{}}}},"/movements":{"get":{"operationId":"movements","parameters":[{"name":"transaction_id","in":"query","required":false,"type":"string"},{"name":"type","in":"query","required":false,"type":"string"},{"name":"account","in":"query","required":false,"type":"string"},{"name":"brand","in":"query","required":false,"type":"string"}],"responses":{"0":{},"200":{}}}},"/transactions":{"get":{"operationId":"transactions","parameters":[{"name":"id","in":"query","required":false,"type":"string"}],"responses":{"200":{}}}},"/transfers":{"post":{"operationId":"transfer","consumes":["application/json"],"parameters":[{"in":"body","name":"body","description":"Request body","required":true}],"responses":{"0":{},"200":{}}}}}}